Refactoring with user_id + new unitary tests
Refactoring with user_id + new unitary tests

+++ src/main/java/com/somedamnmusic/dumb/DumbDatabase.java
@@ -3,9 +3,12 @@ package com.somedamnmusic.dumb;
 import java.util.HashMap;
 import java.util.Map;
 
+import org.apache.commons.lang3.StringUtils;
+
 import com.google.inject.Singleton;
 import com.google.protobuf.ByteString;
 import com.somedamnmusic.apis.DatabaseService;
+import com.somedamnmusic.apis.exception.DatabaseException;
 
 @Singleton
 public class DumbDatabase implements DatabaseService {
@@ -13,11 +16,17 @@ public class DumbDatabase implements DatabaseService {
 	
 	private Map<String, ByteString> database = new HashMap<String, ByteString>();
 
-	public ByteString get(String key) {
+	public ByteString get(String key) throws DatabaseException {
+		if(StringUtils.isBlank(key)) {
+			throw new DatabaseException();
+		}
 		return database.get(key);
 	}
 
-	public void set(String key, ByteString content) {
+	public void set(String key, ByteString content) throws DatabaseException {
+		if(StringUtils.isBlank(key) || content == null) {
+			throw new DatabaseException();
+		}
 		database.put(key, content);
 
 	}
@@ -27,7 +36,10 @@ public class DumbDatabase implements DatabaseService {
 		return "key_"+count;
 	}
 
-	public void remove(String key) {
+	public void remove(String key) throws DatabaseException {
+		if(StringUtils.isBlank(key)) {
+			throw new DatabaseException();
+		}
 		database.remove(key);
 	}
